Index filename, date embedded in filename, and date indexed
This adds three new fields:
  * indexing_date: date the indexer was run
  * input_file_name: basename of the source file from
    multifile_marc4j_reader.rb
  * input_file_date: YYYYMMDD derived from that filename, if possible

Dates are indexed/stored as ints, YYYYMMDD.

The input filename is passed from the reader via an environment
variable (ENV['multifile.filename']). There's no other way for the
reader to communicate with the indexer (which should be fixed
upstream in traject).

Note that the `to_field` calls create closures, so we need to
check the ENV variable within each call, not outside as we
can with "today"'s date.

Also throws in some short (5 records each) sample marc.