Break out configuration/requires for indexers into `settings.rb`
Currently, test-running subsets of the indexers is a pain because
all the configuration, `require` statements, `extend`/`include`
of macros, etc. is at the top of `commong.rb`.

This simply moves all that stuff from the top of `common.rb`
into its own file, `settings.rb`, and adds that new file
to the `mindex_*` scripts.