Use a pushgateway for index job metrics
While sidekiq is itself an offline queue, we're interested in when
IndexIt ran versus when IndexHathi ran, and those specific jobs are
transient. When storing those metrics just in sidekiq, we had to deal
with the inconsistent numbers coming from different replicas, and each
replica would forget its past state on reboot.

This is exactly the sort of thing that the pushgateway was designed for,
but an offline queue (sidekiq) pushing metrics to a pushgateway was
exactly not the sort of thing that Yabeda was designed for. Yabeda saves
a lot of typing but assumes a field of metrics that will change over
time (e.g. number of requests, number of books checked out).

With cronjobs, pushing to a pushgateway works just fine, since the state
will disappear with the cronjob. With sidekiq, the state was persisting
between cronjobs, resulting in weird behavior, especially with multiple
replicas trying to override each other.

By using the prometheus client directly, we can init a new registry and
push metrics with different grouping keys so that nothing overrides
anything unexpected.