Unify method of normalizing text for catalog solr and browse
The `browse_match` fieldType has been changed to mirror
what we're doing when generating the authorities
data, in the hopes that the number of matches reported
in the browse will match the number found when you
click on the link.

In addition to the normal icu normalization stuff,
this removes all punctuation from the analyzed string
in two steps, first turning any of `-:/()[]{}=&` into spaces and
then deleting everything else.