Add detection and use of um-local lc-like subject headings
This adds the concepts of "lc_subject" and "pretender_lc_subject"
so we can at times differentiate from, and at times ignore the
difference betwee, "real" lcsh and our local additions.

The macro now calls "#treated_like_lcsh_fields" to get everything
we're going to use as lcsh, which includes the umich stuff.

Added tests; minor changes to other tests to reflect that we're
treating these local things as lcsh.