change solr fields for remediated subjects
topicAllStr exists in the record so that the full list of searchable
topics is viewable in Solr docs.

topics aren't copied into topicStr anymore because topicStr is now
populated by Subjects::subject_facets.

remediated_lc_subject_display is added so we can have a separate field
in catalog record displays.