fixes identifying lcsh fields
Only some subject fields can indicated lcsh at all. Of those, they need
to have ind2 = 0 and a subfield2 that makes sense. We want to know when
there's a subfield 2 and what it is.

There isn't a straightforward way to pass the doc id at this time, so
we're going to see how common this is at all and if maybe we can
identify these records just by what field it is.